// src/types/graphql-generated/graphql.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};

export type AuthPayload = {
  __typename?: 'AuthPayload';
  token: Maybe<Scalars['String']>;
  userErrors: Array<UserError>;
};

export type ConfirmPayload = {
  __typename?: 'ConfirmPayload';
  confirmed: Maybe<Scalars['Boolean']>;
  userErrors: Array<UserError>;
};

export type CredentialsInput = {
  email: Scalars['String'];
  password: Scalars['String'];
};

export type GetUserPayload = {
  __typename?: 'GetUserPayload';
  user: Maybe<User>;
  userErrors: Array<UserError>;
};
// ...
export type Project = {
  __typename?: 'Project';
  createdAt: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  name: Scalars['String'];
  updatedAt: Maybe<Scalars['String']>;
};
// ...
export type QueryGetUserArgs = {
  id: Scalars['ID'];
};

export type SignupInput = {
  credentials: CredentialsInput;
  name: Scalars['String'];
  passwordConfirm: Scalars['String'];
};

export type Subscription = {
  __typename?: 'Subscription';
  _: Maybe<Scalars['Boolean']>;
};

export type Ticket = {
  __typename?: 'Ticket';
  comment: Maybe<Scalars['String']>;
  createdAt: Maybe<Scalars['String']>;
  description: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  priority: TicketPriority;
  projectId: Scalars['String'];
  references: Maybe<Array<Maybe<Scalars['String']>>>;
  status: TicketStatus;
  title: Scalars['String'];
  updatedAt: Maybe<Scalars['String']>;
  userId: Scalars['String'];
};
// ...
export type User = {
  __typename?: 'User';
  confirmed: Maybe<Scalars['Boolean']>;
  createdAt: Maybe<Scalars['String']>;
  email: Scalars['String'];
  id: Scalars['ID'];
  name: Scalars['String'];
  projects: Maybe<Array<Project>>;
  tickets: Maybe<Array<Ticket>>;
  updatedAt: Maybe<Scalars['String']>;
};

export type UserError = {
  __typename?: 'UserError';
  message: Scalars['String'];
  values: Maybe<Array<Maybe<Scalars['String']>>>;
};